create store.js fix const in app
