+++ src/index.js
@@ -1,21 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import {createStore, applyMiddleware, compose} from 'redux';
+import enhancer from './store';
 import {Provider} from 'react-redux';
+import {createStore} from 'redux';
 import {reducer} from './reducer/reducer';
 
 import App from './components/app/app.jsx';
 import {mapConfig} from './mocks/map-config';
-
-const composeEnhancers =
-  typeof window === `object` &&
-  window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ ?
-    window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__({
-    }) : compose;
-
-const enhancer = composeEnhancers(
-    applyMiddleware()
-);
 
 const store = createStore(reducer, enhancer);
